- Enabling compilation warning for unused and uninitialized variables and sign compare - Removing logging from istgt_create_uctl regarding address (sock=%d, addr=%s, peer=%s) - Changing istgt malloc/free/strdup API to glibc API - Changing memory allocation for cmd(for spec and replica in replication module) from mempool to malloc/free - Disabling mempool allocation for ISTGT.

+++ src/replication_test.c
@@ -607,7 +607,7 @@ again:
 					if (count < 0) {
 						rc = -1;
 						goto error;
-					} else if (count != mgmtio->len) {
+					} else if ((uint64_t)count != mgmtio->len) {
 						REPLICA_ERRLOG("failed to getch mgmt data.. got only %ld bytes out of %lu\n",
 						    count, mgmtio->len);
 						rc = -1;
